// src/Domain/types/dentalChartDetailsStatus.enum.ts
export enum DentalChartDetailsStatus {
  HEALTHY = "HEALTHY",

  CARIES = "CARIES",

  FILLED = "FILLED",

  FRACTURED = "FRACTURED",

  WORN = "WORN",

  MISSING = "MISSING",

  EXTRACTED = "EXTRACTED",

  ROOT_CANAL_TREATED = "ROOT_CANAL_TREATED",

  CROWN = "CROWN",

  IMPLANT = "IMPLANT",

  BRIDGE = "BRIDGE",

  PROSTHETIC = "PROSTHETIC",

  IMPACTED = "IMPACTED",

  MOBILE = "MOBILE",

  INFECTED = "INFECTED",

  ABSCESS = "ABSCESS",

  PERIODONTAL_AFFECTATION = "PERIODONTAL_AFFECTATION",

  SENSITIVITY = "SENSITIVITY",

  DISCOLORATION = "DISCOLORATION",

  DEVELOPMENTAL_ANOMALY = "DEVELOPMENTAL_ANOMALY",

  OTHER = "OTHER"
}